requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
python-dotenv==1.0.0
openai==1.12.0
sentence-transformers==2.2.2
numpy==1.24.3
sqlite3==2.6.0
logging==0.5.1.2
typing==3.7.4.3